# internal/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hypervisor orchestration API for managing staged releases, deployments, and hyperuser access tooling.",
        "title": "OpenHack Hypervisor API",
        "contact": {},
        "version": "25.10.17.4"
    },
    "basePath": "/hypervisor",
    "paths": {
...
        "/hypervisor/env/template": {
            "get": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Env"
                ],
                "summary": "Get template env",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.envTemplateResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Env"
                ],
                "summary": "Update template env",
                "parameters": [
                    {
                        "description": "Template env contents",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.updateEnvTemplateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errmsg._EnvTemplateInvalidRequest"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/hypervisor/releases": {
            "get": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Releases"
                ],
                "summary": "List synced releases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listReleasesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
        "/hypervisor/releases/sync": {
            "post": {
                "security": [
                    {
                        "HyperUserAuth": []
                    }
                ],
                "description": "Triggers a Git pull of the releases repository to refresh available releases for staging.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Hypervisor Releases"
                ],
                "summary": "Sync release metadata",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.StatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errmsg._InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "api.envTemplateResponse": {
            "type": "object",
            "properties": {
                "envText": {
                    "type": "string"
                }
            }
        },
        "api.listReleasesResponse": {
            "type": "object",
            "properties": {
                "releases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Release"
                    }
                }
            }
        },
...
        "api.updateEnvTemplateRequest": {
            "type": "object",
            "properties": {
                "envText": {
                    "type": "string"
                }
            }
        },
...
        "errmsg._EnvTemplateInvalidRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid template env payload"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 400
                }
            }
        },
...
        "errmsg._InternalServerError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "internal server error: \u003cdetails\u003e"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
...
        "models.Release": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sha": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "HyperUserAuth": {
            "description": "Provide the hyperuser bearer token as `Bearer \u003ctoken\u003e`.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}